# backend/src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Controle de Estoque - Logbook",
    "description": "Documentação da API do controle de estoque - Logbook"
  },
  "servers": [
    {
      "url": "http://localhost:3000/"
    }
  ],
  "paths": {
...
    "/api/answers/": {
      "get": {
        "tags": [
          "Answer"
        ],
        "description": "",
        "parameters": [
          {
            "name": "_page",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "_size",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": {}
          }
        ]
      },
      "post": {
        "tags": [
          "Answer"
        ],
        "description": "",
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/answer"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/answer"
              }
            }
          }
        }
      }
    },
    "/api/answers/{_id}": {
      "get": {
        "tags": [
          "Answer"
        ],
        "description": "",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": {}
          }
        ]
      },
      "delete": {
        "tags": [
          "Answer"
        ],
        "description": "",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": {}
          }
        ]
      }
    },
...
    "/api/reagentes/": {
      "get": {
        "tags": [
          "Reagente"
        ],
        "description": "",
        "parameters": [
          {
            "name": "_page",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "_size",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": {}
          }
        ]
      },
      "post": {
        "tags": [
          "Reagente"
        ],
        "description": "",
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/reagente"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/reagente"
              }
            }
          }
        }
      }
    },
    "/api/reagentes/{_id}": {
      "get": {
        "tags": [
          "Reagente"
        ],
        "description": "",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": {}
          }
        ]
      },
      "put": {
        "tags": [
          "Reagente"
        ],
        "description": "",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/reagente"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/reagente"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Reagente"
        ],
        "description": "",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": {}
          }
        ]
      },
      "patch": {
        "tags": [
          "Reagente"
        ],
        "description": "",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": {}
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "InternalServerError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": ""
          },
          "message": {
            "type": "string",
            "example": ""
          }
        },
        "xml": {
          "name": "InternalServerError"
        }
      },
      "reagente": {
        "type": "object",
        "properties": {
          "codigo": {
            "type": "string",
            "example": "123"
          },
          "reagente": {
            "type": "string",
            "example": "abc"
          },
          "quantidade": {
            "type": "number",
            "example": 0
          },
          "dataRecebido": {
            "type": "string",
            "example": "00/00/0000"
          },
          "unidadeMedida": {
            "type": "string",
            "example": "KG"
          },
          "situacao": {
            "type": "string",
            "example": "RECEBIDO"
          },
          "valorUnitario": {
            "type": "number",
            "example": 0
          },
          "valorTotal": {
            "type": "number",
            "example": 0
          },
          "fornecedor": {
            "type": "string",
            "example": "ABC"
          },
          "lote": {
            "type": "string",
            "example": "ABC123"
          },
          "validade": {
            "type": "string",
            "example": "00/00/0000"
          },
          "localizacao": {
            "type": "string",
            "example": "A1"
          },
          "armario": {
            "type": "string",
            "example": "B2"
          },
          "prateleira": {
            "type": "string",
            "example": "C3"
          },
          "solicitante": {
            "type": "string",
            "example": "ABC"
          },
          "limiteMin": {
            "type": "number",
            "example": 0
          },
          "limiteMax": {
            "type": "number",
            "example": 0
          }
        },
        "xml": {
          "name": "reagente"
        }
      },
      "answer": {
        "type": "object",
        "properties": {
          "responsavel": {
            "type": "string",
            "example": "Abc"
          },
          "codigo": {
            "type": "string",
            "example": "123"
          },
          "reagente": {
            "type": "string",
            "example": "ABC"
          },
          "quantidade": {
            "type": "number",
            "example": 0
          },
          "medida": {
            "type": "string",
            "example": "KG"
          },
          "observacao": {
            "type": "string",
            "example": "Utilizado"
          }
        },
        "xml": {
          "name": "answer"
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}